+++ cucumber-core/src/main/java/io/cucumber/core/plugin/Jackson.java
@@ -1,17 +1,17 @@
 package io.cucumber.core.plugin;
 
-import com.fasterxml.jackson.annotation.JsonCreator.Mode;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.databind.cfg.ConstructorDetector;
-import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
-import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
+import io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.annotation.JsonCreator.Mode;
+import io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.annotation.JsonInclude.Include;
+import io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.core.JsonGenerator;
+import io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.databind.DeserializationFeature;
+import io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.databind.ObjectMapper;
+import io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.databind.SerializationFeature;
+import io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.databind.cfg.ConstructorDetector;
+import io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.databind.json.JsonMapper;
+import io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
+import io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 
-import static com.fasterxml.jackson.annotation.JsonInclude.Value.construct;
+import static io.cucumber.messages.ndjson.internal.com.fasterxml.jackson.annotation.JsonInclude.Value.construct;
 
 final class Jackson {
     public static final ObjectMapper OBJECT_MAPPER = JsonMapper.builder()